Fixed the spawn directions
Now made it so that the player spawns facing the centre of the map using
the new rotation feature added to the map. Also press L to bring up the
debug menu, currently only used to spit out the most recent spawn point
in the most awful way possible in code.